PokeAByte.Domain. Moved extension classes to their own folder. Some cleanup.
- Documentation for  NormalizeMemoryAddresses(), as I was initially mistaken about it's purpose :)
- Added ParseHexAddress() extension method to avoid double-parsing strings.
- In ByteExtensions, made 'EndianTypes' params non-nullable.
- Removed various unused extension methods.
